migrate javascript to typescript
